fix: move abstract class to commons package and codestyle fixes
